refactor: move the TxIdComputable trait to strata-primitives
